MAKE `CloudStorage` util class for housing properties, getters & setters
ADD extension of `CloudStorage` to `S3`.
ADD protected `uploadStream()` method to `S3` for executing streamed uploads.
